bound

+++ merge/source/basis/bases_2D/basis_dubiner_2D.cpp
@@ -2,14 +2,14 @@
 
 namespace Basis{
 Array2D<double> Dubiner_2D::get_phi(int p, const std::vector<Point<2>>& coordinates) {
-	size_t n_pts = coordinates.size();
+	int n_pts = coordinates.size();
 	
 	Array2D<double> phi((p + 1)*(p + 2) / 2);
 	
 	std::vector<double> n1(n_pts);
 	std::vector<double> n2(n_pts);
 
-	for (size_t i = 0; i < n_pts; i++) {
+	for (int i = 0; i < n_pts; i++) {
 		n1[i] = 2 * (1 + coordinates[i][Z1]) / (1 - coordinates[i][Z2]) - 1;
 		n2[i] = coordinates[i][Z2];
 
@@ -28,14 +28,14 @@ Array2D<double> Dubiner_2D::get_phi(int p, const std::vector<Point<2>>& coordina
 }
 
 Array3D<double> Dubiner_2D::get_dphi(int p, const std::vector<Point<2>>& coordinates) {
-	size_t n_pts = coordinates.size();
+	int n_pts = coordinates.size();
 
 	Array3D<double> dphi((p + 1)*(p + 2) / 2);
 
 	std::vector<double> n1(n_pts);
 	std::vector<double> n2(n_pts);
 
-	for (size_t i = 0; i < n_pts; i++) {
+	for (int i = 0; i < n_pts; i++) {
 		n1[i] = 2 * (1 + coordinates[i][Z1]) / (1 - coordinates[i][Z2]) - 1;
 		n2[i] = coordinates[i][Z2];
 	}
@@ -74,7 +74,7 @@ void Dubiner_2D::basis_test(int p, const Array2D<double>& phi_area, const std::p
 		M[i].reserve(number_bf);
 		for (int j = 0; j < number_bf; j++) {
 			M[i].push_back(0);
-			for (size_t k = 0; k < integration_rule.first.size(); k++) {
+			for (int k = 0; k < integration_rule.first.size(); k++) {
 				M[i][j] = M[i][j] + integration_rule.first[k] * phi_area[i][k] * phi_area[j][k];
 			}
 		}
@@ -107,14 +107,14 @@ void Dubiner_2D::basis_test(int p, const Array2D<double>& phi_area, const std::p
 }
 
 std::vector<double> Dubiner_2D::dubiner_2d_phi(int p, int q, const std::vector<double>& n1, const std::vector<double>& n2) {
-	size_t n_pts = n1.size(); //CHECK IF n1.size() = n2.size()
+	int n_pts = n1.size(); //CHECK IF n1.size() = n2.size()
 	
 	std::vector<double> phi(n_pts);
 	
 	std::vector<double> psi_p(jacobi_polynomial(p, 0, 0, n1));
 	std::vector<double> psi_pq(jacobi_polynomial(q, 2 * p + 1, 0, n2));
 
-    for (size_t i = 0; i < n_pts; i++) {
+    for (int i = 0; i < n_pts; i++) {
 		phi[i] = psi_p[i] * pow((1 - n2[i]) / 2, p)*psi_pq[i];
 
 		if (n1[i] == NAN) { //value of Dubiner polynomial at singular point (-1,1)
@@ -131,7 +131,7 @@ std::vector<double> Dubiner_2D::dubiner_2d_phi(int p, int q, const std::vector<d
 }
 
 Array2D<double> Dubiner_2D::dubiner_2d_dphi(int p, int q, const std::vector<double>& n1, const std::vector<double>& n2) {
-	size_t n_pts = n1.size(); //CHECK IF n1.size() = n2.size()
+	int n_pts = n1.size(); //CHECK IF n1.size() = n2.size()
 	
 	Array2D<double> dphi_d(2);
 	dphi_d[Z1].reserve(n_pts);
@@ -143,7 +143,7 @@ Array2D<double> Dubiner_2D::dubiner_2d_dphi(int p, int q, const std::vector<doub
 	std::vector<double> dpsi_p_dn1(jacobi_polynomial_derivative(p, 0, 0, n1));
 	std::vector<double> dpsi_pq_dn2(jacobi_polynomial_derivative(q, 2 * p + 1, 0, n2));
 
-    for (size_t i = 0; i < n_pts; i++) {
+    for (int i = 0; i < n_pts; i++) {
         dphi_d[Z1].push_back((2.0 / (1.0 - n2[i])) * dpsi_p_dn1[i] * pow((1.0 - n2[i]) / 2.0, p)*psi_pq[i]);
 
         dphi_d[Z2].push_back(((1 + n1[i]) / (1.0 - n2[i])) * dpsi_p_dn1[i] * pow((1.0 - n2[i]) / 2.0, p)*psi_pq[i] +
